% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{avg_dist_func}
\alias{avg_dist_func}
\title{Create average distance matrix}
\usage{
avg_dist_func(dist_mat, cohortid.vec)
}
\arguments{
\item{dist_mat}{distance matrix}

\item{cohortid.vec}{a numeric vector of cohort/group id for all observations e.g., \code{rep(1:4,each=100)}}
}
\description{
Convert individual-level distance matrix to average distance matrix at cohort level
}
