% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{BGcompare}
\alias{BGcompare}
\title{Wrapper function of BG test}
\usage{
BGcompare(
  X,
  Y,
  BG.method = c("asymptotic", "permutation", "automatic"),
  .n_perm,
  N_auto
)
}
\arguments{
\item{X, Y}{two dataframes for comparisons}

\item{BG.method}{string, "asymptotic","permutation","automatic"}

\item{.n_perm}{number of permuation if method is permutation}

\item{N_auto}{integer, the sample size for automatically choosing between asymptotic and permutation, below is permutation, above is asymptotic}
}
\description{
This is a wrapper function of Biswas and Ghosh multivariate test implemented in \code{SHT::eqdist.2014BG}
}
