% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{df}
\alias{df}
\title{Simulated Dataset with three clusters}
\format{
An object of class data.frame with 1000 rows and 17 columns:
\describe{
\item{X1-X15}{numeric, simulated continuous variables from multivariate joint distributions}
\item{cohortid}{numeric, cohort id}
\item{distribution}{numeric, true distribution indicator}
}
}
\usage{
data(df)
}
\description{
Simulated multivariate continuous samples from three cohort clusters (1-3 ,4-6, 7-10) with different mean vectors
}
\keyword{datasets}
