% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_na_level}
\alias{find_na_level}
\title{Examine missing proportion by group}
\usage{
find_na_level(df, clusterid)
}
\arguments{
\item{df}{data.frame}

\item{clusterid}{string, the name of cohort id indicator in the dataframe}
}
\description{
A wrapper function for examining missingness level by cohort/cluster
}
