\name{print.relate}
\alias{print.relate}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{prints the results for relate }
\description{prints the results for relate }
\usage{
\method{print}{relate}(x,...)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{x}{an relate object from the relate function}
  \item{...}{arguments passed to print.power.htest}
}
\keyword{manip}

